+++ kafka/config.go
@@ -16,37 +16,33 @@ type ConsumerConfig struct {
 }
 
 func (c ConsumerConfig) toConfigMap() kafka.ConfigMap {
-	autoCommitInterval := 15000
-	debugLevel := "consumer"
-	autoOffsetReset := "earliest"
-	partitionAssignment := "range,roundrobin"
 
-	if c.AutoCommitInterval > 0 {
-		autoCommitInterval = c.AutoCommitInterval
+	kafkaConfMap := kafka.ConfigMap{
+		"bootstrap.servers":        c.BootstrapServers,
+		"group.id":                 c.GroupID,
+		"enable.auto.commit":       true,
+		"debug":                    "consumer",
+		"go.logs.channel.enable":   true,
+		"enable.auto.offset.store": false,
+		"auto.commit.interval.ms":  15000,
+		"auto.offset.reset":        "earliest",
 	}
-	if c.DebugLevel != "" {
-		debugLevel = c.DebugLevel
-	}
-
 	if c.AutoOffsetReset != "" {
-		autoOffsetReset = c.AutoOffsetReset
+		kafkaConfMap["auto.offset.reset"] = c.AutoOffsetReset
 	}
 
 	if c.PartitionAssignment != "" {
-		partitionAssignment = c.PartitionAssignment
+		kafkaConfMap["partition.assignment.strategy"] = c.PartitionAssignment
 	}
 
-	kafkaConfMap := kafka.ConfigMap{
-		"bootstrap.servers":             c.BootstrapServers,
-		"group.id":                      c.GroupID,
-		"auto.offset.reset":             autoOffsetReset,
-		"enable.auto.commit":            true,
-		"auto.commit.interval.ms":       autoCommitInterval,
-		"debug":                         debugLevel,
-		"go.logs.channel.enable":        true,
-		"enable.auto.offset.store":      false,
-		"partition.assignment.strategy": partitionAssignment,
+	if c.DebugLevel != "" {
+		kafkaConfMap["debug"] = c.DebugLevel
 	}
+
+	if c.AutoCommitInterval > 0 {
+		kafkaConfMap["auto.commit.interval.ms"] = c.AutoCommitInterval
+	}
+
 	return kafkaConfMap
 }