+++ CHANGELOG.md
@@ -1,9 +1,20 @@
+
 # Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.6.0] 2022-02-13
+# Changes
+
+- changes internal retry structure for RabbitMQ
+
+# Breaking Changes
+
+- The RabbitMQ `QueueConfig` now uses `QueueKey` instead of `QueueName`
+
 
 ## [1.5.1] 2022-01-28